Update greenChocolate.html
added a hidden feature for video removal. now, right click the text (those "white on chocolate" below) can remove the respective video from the webpage. with this hidden feature, even if the green spot is unavailable (as when greenChocolate is loaded from an html file), user still has a way to remove the video.