Update greenChocolate.html
added two features: (1) autosave, which will make a copy of the id's of the videos currently on the playground after a drop video event happened or a remove video action taken. and on every start of a session (load the page/html or refresh the page/html), the autosaved copy will be checked. and if anything exists there, the system will ask if you want to bring it back. and please be aware that since the autosaved copy is kept in `localStorage`, it's domain oriented. essentially, different domains can have different set of autosaved copy. (2) added a click event handler for the green board, so that you can click it away in case of the green board didn't go away automatically somehow.