Update greenChocolate.html
added two behaviors: (1) after a drag event has caused the appearance of the green board, the script will see if at least a part of the green board is visible in the viewport. and in case of "none of it is visible", the viewport will be scrolled to the top (to show the green board) automatically. and (2) after a video is inserted due to a drop event, the viewport will be scrolled to the top (to show the newly added video) automatically.