Update README.md
my recently added operation notes (regarding how to add recommended videos) was inaccurate. my observation thus conclusion was drawn on different scenarios on different platforms. i wasn't aware of it, and presumed they were under the same scenario on both platforms. my misunderstanding produced a misleading operation notes, and i'm sorry. so, this is for the adjustment of the notes.