Update greenChocolate.html
added some cosmetic features which are not for new functionalities but for improving the user interface.

(1) added the visual effect that when mouse hovers, (i) the green transparent spot and (ii) the "jump to the top" icon, they will become "solid" instead of transparent.

(2) when the video text (the white on chocolate below the video) is being mouse hovered, a "remove" (red on green) button will pop up at the far end of the text. you may click it to remove the video. so, with this new addition, you may avoid to use the rightclicking (even though it's preserved as muscle memory of it might have already been developed) hidden feature at all.