Update greenChocolate.html
added some cosmetic features which are not for new functionalities but for improving the user interface.

(1) added the visual effect that when mouse hovers, (i) the green transparent spot and (ii) the "jump to the top" icon, they will become "solid" instead of transparent.

(2) there has been two "hidden" features, which are triggered by rightclick, around for some time. and they are somewhat contradictory as (a) rightclicking the video text (the white on chocolate below the video) to remove the video; and (b) rightclicking the green transparent spot to play the video plus the logging of the play. although user is less likely to use the rightclicking of the video text if the green transparent spot is present (because clicking the green transparent spot to remove the video is much easier and intuitive by all means), it still somewhat an inconsistency in the sense when considering rightclicking itself as a category of functionality. i did consider to swap the functionality between the clicking and rightclicking of the green transparent spot so to make rightclicking all about removing. but then it involves a change of muscle memory which is rather expensive. on the other hand, i also reluctant to add one more permanent transparent spot for "play+log". so, i end up settle with this additional user interface which toggles the show/hide of a youtube play button (which can be clicked to trigger the "play+log") whenever the mouse hovers the green transparent spot. hence, you don't have to rightclicking (even though it's preserved as muscle memory of it might have also been developed already) for "play+log", you can just click the popped up transparent youtube play button for it. besides, when the video text is being mouse hovered, a "remove" (red on green) button will pop up at the far end of the text. so, with all these new additions, you may avoid to use rightclicking (even though, again, it's preserved) at all.

(3) made a minor adjustment to the `sel_anchNode()` function.