Update greenChocolate.html
added the import/export (of youtube video id's) feature. it could be useful when you want to transfer the unfinished videos list from one browser to another. for instance, say you've picked a list of youtube videos onto greenChocolate playground, and before you have time to finish them all, you have to go out for something that you may want to export the list of the remaining videos and import it into your greenChocolate playground on your smartphone's browser so that you can watch them whenever possible while you are on the road.