Update greenChocolate.html
added a handy vertical slider for you to change the video playback speed. it has 14 levels from the slowest at 0.25 to the fastest at 3.75 (and up/down per every 0.25) while default at 1 (normal speed). you may show/hide it by mouse hover the green transparent spot. you can change its value by click, or drag, or scroll. click or drag is pretty intuitive, nothing need to talk about. for scroll though, you need to know the effective area where your mouse pointer is. it's not as narrow as the slider is, the effective area is actually all the way extended to the right edge of the player window from the slider.