Update greenChocolate.html
added a feature to automatically set "the best" speed for the youtube video that is starting to play on the greenChocolate playground. you have to define "the best" playback speeds for different channels if you want them to be playing in different default speed than 1 (the normal speed). use the hotkey `alt+s` to edit the setting.