Update greenChocolate.html
there is a ruler (sort of) next to the vertical slider as an auxiliary visual cue for playback speed tuning, that is, those 3 little triangles marking for the speed of one, two and three. while the slider can be resized automatically as the window size is changed (for instance, rotating the screen between portrait and landscape), the ruler didn't resize accordingly. so, this is the replenishment for the omission.