Update greenChocolate.html
added a feature to enable user to adjust the playback speed by touch operation. just start the touch from the transparent green spot and while your finger is still in touch with the screen, slide it somewhere (horizontally, vertically, diagonally, whatever you like, it doesn't matter) and you will see a number representing your target playback speed is shown in the center of the display and is changing (increasing/decreasing) as your finger sliding back and forth (relative to the transparent green spot). and when your target speed is shown, leave your finger off the screen. in turn, the playback speed will be set accordingly.